[TRUNK-13072] Add num_files, num_tests, repo_head_sha_short to meta.json (#131)
* add num_files, num_tests, repo_head_sha_short to meta.json

* fix type error in context-py test

* move import

* pr feedback
