+++ cli/src/lib.rs
@@ -5,5 +5,6 @@ pub mod constants;
 pub mod runner;
 pub mod scanner;
 pub mod types;
+pub mod upload;
 pub mod utils;
 pub mod validate;
