+++ cli/src/types.rs
@@ -117,6 +117,8 @@ pub struct BundleMeta {
     pub bundle_upload_id: String,
     pub tags: Vec<CustomTag>,
     pub file_sets: Vec<FileSet>,
+    pub num_files: usize,
+    pub num_tests: usize,
     pub envs: std::collections::HashMap<String, String>,
     pub upload_time_epoch: u64,
     pub test_command: Option<String>,
@@ -232,6 +234,7 @@ mod tests {
             repo_root: "repo_root".to_string(),
             repo_url: "repo_url".to_string(),
             repo_head_sha: "repo_head_sha".to_string(),
+            repo_head_sha_short: "repo_head_sha_short".to_string(),
             repo_head_branch: "repo_head_branch".to_string(),
             repo_head_commit_epoch: 1724102768,
             repo_head_commit_message: "repo_head_commit_message".to_string(),
